--- project/api/admin/__init__.py
@@ -1,7 +1,0 @@
-from .movie import MovieAdmin
-from .video import VideoAdmin
-
-__all__ = (
-    'MovieAdmin',
-    'VideoAdmin',
-)
