--- project/api/admin/movie.py
@@ -1,10 +1,0 @@
-from django.contrib import admin
-
-from ..models import Movie
-
-
-@admin.register(Movie)
-class MovieAdmin(admin.ModelAdmin):
-    list_display = ('id', 'title', 'image_url', 'link')
-    search_fields = ('title',)
-    empty_value_display = '-пусто-'
